Create School_Room_Finder.py
Your school is organising an open evening event where prospective parents will be able to visit the different classrooms and departments to find out more about the school and meet with the teachers. Parents have been given a list of rooms and activities going on during this open evening. The school will utilise student leaders to assist parents in navigating between the different buildings. The school would also like to develop an App for parents to use on their phones to help them locate a room.